Make id generation customizable (#203)
Branca/folium uses random numbers to generate ids. This has some
disadvantages. E.g. in the streamlit-folium plugin they have to
replace the random ids with predictable ids for comparison between
streamlit runs.

Also, in the folium tests, we have to normalize the variable names
to compare between expected and actual results. This change in Branca
makes it possible to override or customize id generation. This allows
client libraries to experiment with new id generation schemes.